[CHORE] social login ui
