+++ client/src/Router.js
@@ -14,6 +14,7 @@ import Login from "./pages/Login";
 import SignUp from "./pages/SignUp";
 import Search from "./pages/Search";
 import CalculetList from "./pages/CalculetList";
+import WriteUserInfo from "./pages/WriteUserInfo";
 
 // google-analytics
 import RouteChangeTracker from "./components/google-analytics/RouteChangeTracker";
@@ -33,6 +34,7 @@ const AppRouter = () => {
         />
         <Route path={URL.LOGIN} element={<Login />} />
         <Route path={URL.SIGN_UP} element={<SignUp />} />
+        <Route path={URL.WRITE_USER_INFO} element={<WriteUserInfo />} />
         <Route path={URL.SEARCH} element={<Search />} />
         <Route path={URL.CALCULET_LIST} element={<CalculetList />} />
       </Routes>
