+++ client/src/components/sign-up/EmailForm.js
@@ -5,23 +5,21 @@ import {
 import { OPTIONS_EMAIL_ADDRESS } from "./constants";
 import { InformBox } from "../register/WriteInform";
 /**
- * 회원가입 1번째 정보입력창
+ * 회원가입 - 이메일로 가입 입력창
  *
  * @param {string, string, string, string, string, string, funtion, funtion, funtion, funtion, funtion, funtion} props
  * address : 주소
  * writtenDomain : 직접 입력한 도메인 값
  * selectedDomain : select 박스에서 선택한 도메인 값
- * userName : 닉네임
  * pw : 비밀번호
  * pwConfirmation : 비밀번호 확인
  * changeAddress : 주소 change 함수
  * changeDomain : 직접 입력한 도메인 값 change 함수
  * changeSelectedDomain : select 박스에서 선택한 도메인 값 change 함수
- * changeUserName : 닉네임 change 함수
  * changePw : 비밀번호 change 함수
  * changePwConfirmation : 비밀번호 확인 change 함수
  */
-function WriteInformFirst(props) {
+function EmailForm(props) {
   return (
     <InformBox>
       <ExplanationEmail
@@ -35,15 +33,6 @@ function WriteInformFirst(props) {
         onChangeDomain={props.changeDomain}
         emailAddressOptions={OPTIONS_EMAIL_ADDRESS}
         onChangeSelectedDomain={props.changeSelectedDomain}
-      />
-      <ExplanationInputBox
-        isLine={true}
-        ratioLeft="1"
-        ratioRight="3.5"
-        explanation="닉네임"
-        placeholder="닉네임을 입력하세요."
-        defaultValue={props.userName}
-        onChange={props.changeUserName}
       />
       <ExplanationInputBox
         type="password"
@@ -68,4 +57,4 @@ function WriteInformFirst(props) {
     </InformBox>
   );
 }
-export default WriteInformFirst;
+export default EmailForm;
