+++ client/src/pages/SignUp.js
@@ -1,29 +1,27 @@
 import { useEffect, useState } from "react";
 import styled from "styled-components";
+import styles from "../components/styles";
 import { BoxBorder } from "../components/atom-components/BoxBorder";
 import { StyledImg } from "../components/atom-components/BoxIcon";
 import BoxTitle from "../components/atom-components/BoxTitle";
-import { BtnIndigo } from "../components/atom-components/ButtonTemplate";
+import {
+  BtnIndigo,
+  StyledIcon,
+} from "../components/atom-components/ButtonTemplate";
 import {
   ContentLayout,
   FlexColumnLayout,
+  FlexRowLayout,
   White300Layout,
 } from "../components/Layout";
-import {
-  OPTIONS_SEX,
-  OPTIONS_YEAR,
-  OPTIONS_MONTH,
-} from "../components/sign-up/constants";
-import fillOptionsList from "../components/sign-up/fillOptionsList";
 import { OPTIONS_EMAIL_ADDRESS } from "../components/sign-up/constants";
-import ProfileChange from "../components/sign-up/ProfileChange";
-import WriteInformFirst from "../components/sign-up/WriteInformFirst";
-import WriteInformSecond from "../components/sign-up/WriteInformSecond";
 import ActGuide from "../components/sign-up/ActGuide";
 import WarningGuide from "../components/global-components/WarningGuide";
 import useInput from "../hooks/useInput";
-import signUpUser from "../user-actions/SignUpUser";
 import { useNavigate } from "react-router-dom";
+import EmailForm from "../components/sign-up/EmailForm";
+import SmallTitle from "../components/global-components/SmallTitle";
+
 /**
  * 흰색 뒷 배경
  */
@@ -50,51 +48,39 @@ const WrapperStretch = styled(FlexColumnLayout)`
   width: 100%;
 `;
 
+const Title = styled.div`
+  ${styles.sytleText.text200};
+`;
+
+const Logo = styled.img`
+  width: 24px;
+  height: 24px;
+  align-items: center;
+  justify-content: center;
+`;
+
 /**
  * 회원가입 페이지
  */
 function SignUp() {
   /**
-   * 프로필 사진 profileImg - type : Blob
-   *
    * 이메일 email -> address, writtenDomain, selectedDomain
-   * 닉네임 userName
    * 비밀번호 pw
    * 비밀번호 확인 pwConfirmation
-   *
-   * 성별 sex
-   * 생년월일 birthdate -> year, month, date
-   * 직업 job
-   * 자기소개 문구 bio
    */
-  const [profileImg, setProfileImg] = useState("/img/defaultProfile.png");
-
   const address = useInput("");
   const [domain, setDomain] = useState("");
   const [writtenDomain, setwrittenDomain] = useState("");
   const [selectedDomain, setSelectedDomain] = useState("");
 
-  const userName = useInput("");
   const pw = useInput("");
   const pwConfirmation = useInput("");
 
-  const [sex, setSex] = useState("");
-
-  const [year, setYear] = useState("");
-  const [month, setMonth] = useState("");
-  const [date, setDate] = useState("");
-
-  const job = useInput("");
-  const bio = useInput("");
-
   // 주의 문구 여부 : 비밀번호 유효성 검사 / 비밀번호 & 비밀번호 확인 비교
   const [warningPw, setWarningPw] = useState("");
+
   // 주의 문구 여부: 다 입력되었는지 여부 & 요청 정보 오류
   const [warningAll, setWarningAll] = useState("");
-
-  // 년도와 월로 일수 구하기
-  const [dateEnd, setDateEnd] = useState(1);
-  const [dates, setDates] = useState(null);
 
   // 라우터 역할 네이게이션
   const navigate = useNavigate();
@@ -104,14 +90,6 @@ function SignUp() {
 
   // 선택되거나 입력된 도메인 갱신
   useEffect(() => setDomain(writtenDomain), [writtenDomain]);
-
-  // 년도에 따른 월의 마지막 날 계산 & 일수 구하기 (년도와 월을 둘 다 선택해야 갱신)
-  useEffect(() => {
-    if (year && month) {
-      setDateEnd(Number(new Date(Number(year), Number(month), 0).getDate()));
-      setDates(fillOptionsList(1, dateEnd));
-    }
-  }, [year, month, dateEnd]);
 
   // 비밀번호 유효성 검사 (10자 이상, 알파벳과 특수문자 포함)
   // 비밀번호 & 비밀번호 확인 같은지 검사
@@ -150,20 +128,6 @@ function SignUp() {
   }
 
   /**
-   * 인자로 넘어온 정보에 대한 change 함수
-   * - value에 먼저 접근한 후, value에 맞는 name을 찾아서 저장
-   * @param {*} event
-   * event : 이벤트
-   * optionData : 옵션 데이터
-   * setData : change할 state
-   */
-  function changeData(event, optionData, setData) {
-    const targetValue = event.target.value;
-    const option = optionData.filter((x) => x.value === targetValue);
-    setData(option[0].name);
-  }
-
-  /**
    * 폼 제출
    * - 입력된 비밀번호와 비밀번호 확인에 따른 경고 안내문 & 회원가입 성공
    */
@@ -172,48 +136,12 @@ function SignUp() {
     // 비밀번호 & 비밀번호 확인 비교
     if (warningPw) return;
     // 다 입력했는지 확인
-    if (
-      !address.value ||
-      !domain ||
-      !userName.value ||
-      !pw.value ||
-      !pwConfirmation.value ||
-      !sex ||
-      !year ||
-      !month ||
-      !date ||
-      !job.value ||
-      !bio.value
-    ) {
+    if (!address.value || !domain || !pw.value || !pwConfirmation.value) {
       setWarningAll("모든 사항을 입력해 주세요.");
       return;
     } else setWarningAll("");
 
-    // DB 데이터 타입에 맞게 처리
-    const sexDb = sex === "여자" ? "F" : "M";
-    const emailDb = address.value + "@" + domain;
-    const birthdateDb = year + "-" + month + "-" + date;
-
-    // 서버에 보낼 정보 => body
-    let body = {
-      email: emailDb,
-      userName: userName.value,
-      profileImg: profileImg,
-      bio: bio.value,
-      sex: sexDb,
-      birthdate: birthdateDb,
-      pw: pw.value,
-      job: job.value,
-    };
-
-    // 서버에 요청
-    const request = signUpUser(body);
-    request.then((res) => {
-      // 회원 가입 실패
-      if (res.message) setWarningAll("올바른 정보를 입력해 주세요.");
-      // 회원 가입 성공
-      else if (res.location) navigate("/login");
-    });
+    // firebase 통한 이메일 회원가입 진행
   }
   return (
     <>
@@ -222,61 +150,42 @@ function SignUp() {
         <StyledImg src="/ATCLogoBlueImgText.png" width="214px" />
         <BoxBorder gap="20px">
           <BoxTitle content="회원가입" />
-          <ProfileChange
-            profileImg={profileImg}
-            setProfileImg={setProfileImg}
-          />
           <WrapperStretch gap="10px">
-            <WriteInformFirst
+            <EmailForm
               address={address.value}
               writtenDomain={writtenDomain}
               selectedDomain={selectedDomain}
-              userName={userName.value}
               pw={pw.value}
               pwConfirmation={pwConfirmation.value}
               changeAddress={address.onChange}
               changeDomain={(event) => setwrittenDomain(event.target.value)}
               changeSelectedDomain={changeSelectedDomain}
-              changeUserName={userName.onChange}
               changePw={pw.onChange}
               changePwConfirmation={pwConfirmation.onChange}
             />
             {warningPw && <WarningGuide content={warningPw} />}
-          </WrapperStretch>
-          <WrapperStretch gap="10px">
-            <WriteInformSecond
-              sex={sex}
-              year={year}
-              month={month}
-              date={date}
-              dates={dates}
-              job={job.value}
-              bio={bio.value}
-              changeSex={(event) => changeData(event, OPTIONS_SEX, setSex)}
-              changeYear={(event) => {
-                changeData(event, OPTIONS_YEAR, setYear);
-                setDate(null);
-              }}
-              changeMonth={(event) => {
-                changeData(event, OPTIONS_MONTH, setMonth);
-                setDate(null);
-              }}
-              changeDate={(event) => changeData(event, dates, setDate)}
-              changeJob={job.onChange}
-              changeBio={bio.onChange}
-            />
             {warningAll && <WarningGuide content={warningAll} />}
           </WrapperStretch>
           <WrapperStretch>
             <BtnIndigo text="가입하기" onClick={onSubmitHandler} />
           </WrapperStretch>
-        </BoxBorder>
-        <BoxBorder>
           <ActGuide
             guide="이미 계정이 있으신가요?"
             lead="로그인하기"
             onClick={() => navigate("/login")}
           />
+        </BoxBorder>
+        <BoxBorder>
+          <FlexRowLayout gap="20px">
+            <Logo src="/img/googleLogo.png" />
+            <Title>Google 계정으로 가입 하기</Title>
+          </FlexRowLayout>
+        </BoxBorder>
+        <BoxBorder>
+          <FlexRowLayout gap="20px">
+            <Logo src="/img/githubLogo.png" />
+            <Title>Github 계정으로 가입 하기</Title>
+          </FlexRowLayout>
         </BoxBorder>
       </WrapperPad>
     </>
